requirements.txt
gradio
torch
torchvision
einops
numpy
pillow